+++ Exareme-Docker/src/mip-algorithms/CART/init/1/global.py
@@ -11,6 +11,7 @@ sys.path.append(path.dirname(path.dirname(path.dirname(path.dirname(path.abspath
 
 from algorithm_utils import StateData, ExaremeError
 from cart_lib import CartInit_Loc2Glob_TD, Cart_Glob2Loc_TD
+from cart_steps import cart_init_1_global
 
 def main():
     # Parse arguments
@@ -27,8 +28,9 @@ def main():
     args_X, args_Y, CategoricalVariables = CartInit_Loc2Glob_TD.load(local_dbs).get_data()
 
     # Run algorithm global step
-    globalTree = None
-    activePaths = None
+    #globalTree = None
+    #activePaths = None
+    globalTree, activePaths = cart_init_1_global()
 
     # Save global state
     global_state = StateData(   stepsNo  = 0 ,
